feat: working on a get_matching! macro
working on a get_matching! macro, happy with how it
works so far, but hacked it to work by hard coding the
test struct 'TesterOpt' into it to get it to work. might
need to rework to get this to work because i'm not sure of
how to determine the type of the object being passed into
the macro, let alone creating a new ident +Opt at the end.

added default to the struct, so you don't have to include
all the parameters. added this to the derive macro too.